+++ src/helpers/DynamicMeta.php
@@ -58,7 +58,7 @@ class DynamicMeta
      * use nystudio107\seomatic\events\AddDynamicMetaEvent;
      * use nystudio107\seomatic\helpers\DynamicMeta;
      * use yii\base\Event;
-     * Event::on(DynamicMeta::class, DynamicMeta::EVENT_INCLUDE_CONTAINER, function(AddDynamicMetaEvent $e) {
+     * Event::on(DynamicMeta::class, DynamicMeta::EVENT_ADD_DYNAMIC_META, function(AddDynamicMetaEvent $e) {
      *     // Add whatever dynamic meta items to the containers as you like
      * });
      * ```